[#845] Split components across multiple crates
