+++ iceoryx2-tunnels/core/src/tunnel/tunnel.rs
@@ -14,7 +14,7 @@ use std::collections::{HashMap, HashSet};
 
 use crate::tunnel::ports::Ports;
 use crate::tunnel::{discovery, ports};
-use crate::{Discovery, Relay, RelayBuilder, RelayFactory, Transport};
+use iceoryx2_tunnels_traits::{Discovery, Relay, RelayBuilder, RelayFactory, Transport};
 
 use iceoryx2::node::{Node, NodeBuilder, NodeCreationFailure};
 use iceoryx2::prelude::PortFactory;
