+++ iceoryx2-tunnels/old/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
-name = "iceoryx2-tunnels-transport-zenoh"
-description = "iceoryx2: zenoh transport for tunnels"
+name = "iceoryx2-tunnels-zenoh-old"
+description = "iceoryx2: tunnel over zenoh"
 categories = { workspace = true }
 edition = { workspace = true }
 homepage = { workspace = true }
@@ -12,11 +12,18 @@ rust-version = { workspace = true }
 version = { workspace = true }
 
 [lib]
-name = "iceoryx2_tunnels_transport_zenoh"
+name = "iceoryx2_tunnels_zenoh_old"
 path = "src/lib.rs"
 
 [dependencies]
 iceoryx2 = { workspace = true }
-iceoryx2-tunnels-core = { workspace = true }
 iceoryx2-bb-log = { workspace = true }
+iceoryx2-services-discovery = { workspace = true }
+
+serde_json = { workspace = true }
 zenoh = { workspace = true }
+
+[dev-dependencies]
+iceoryx2-bb-testing = { workspace = true }
+iceoryx2-bb-posix = { workspace = true }
+generic-tests = { workspace = true }
