+++ iceoryx2-tunnels/old/src/lib.rs
@@ -10,6 +10,14 @@
 //
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 
-mod tunnel;
+pub(crate) mod channel;
+pub(crate) use channel::*;
 
+pub(crate) mod discovery;
+
+pub mod keys;
+
+pub(crate) mod middleware;
+
+mod tunnel;
 pub use tunnel::*;
